Remove complex disabling of ACPI for non-ACPI targets
The disabling of ACPI for non-ACPI targets is rather too complex
and leads to linking issues because of the workarounds to define
objects that get referenced in the link stage but are never used.
Simplfy the build by building in the APCICA core but never calling
it for non-ACPI targets. This removes the need for the VOID_FUNC
workaround stubs in src/libfwtsacpica/fwts_acpica.c

Signed-off-by: Colin Ian King <colin.king@canonical.com>